+++ np.Templating/src/Templating.js
@@ -55,13 +55,13 @@ export async function migrateQuickNotes() {
     // console.log('template: ' + quickNote.template)
     // console.log('title: ' + quickNote.title)
     // console.log('folder: ' + quickNote.folder)
-    const templateFilename = `🗒 Quick Notes/${quickNote.label}`
+    const templateFilename = `Test/🗒 Quick Notes/${quickNote.label}`
     const templateData: ?TNote = await getOrMakeNote(quickNote.template, '📋 Templates')
 
     let title = quickNote.title
     title = title.replace('{{meetingName}}', '<%- meetingName %>')
     title = title.replace('{{MeetingName}}', '<%- meetingName %>')
-    // title = title.replace('{{date8601()}}', '<%- date.now() %>')
+    title = title.replace('{{date8601()}}', '<%- date8601() %>')
     title = title.replace("{{weekDates({format:'yyyy-MM-dd'})}}", "<%- date.startOfWeek('ddd YYYY-MM-DD',null,1) %>  - <%- date.endOfWeek('ddd YYYY-MM-DD',null,1) %>")
     const metaData = {
       newNoteTitle: title,
@@ -209,10 +209,10 @@ export async function templateQuote(): Promise<string> {
   }
 }
 
-export async function templateMigration(silent: boolean = false): Promise<void> {
+export async function migrateTemplates(silent: boolean = false): Promise<void> {
   try {
     const templateFolder = '📋 Templates'
-    const newTemplateFolder: string = '@Templates' // NotePlan.environment.templateFolder
+    const newTemplateFolder: string = '@Templates/Test' // NotePlan.environment.templateFolder
 
     const templateNotes = DataStore.projectNotes.filter((n) => n.filename?.startsWith(templateFolder)).filter((n) => !n.title?.startsWith('_configuration'))
     const newTemplates = DataStore.projectNotes.filter((n) => n.filename?.startsWith(newTemplateFolder)).filter((n) => !n.title?.startsWith('_configuration'))
@@ -258,6 +258,9 @@ export async function templateMigration(silent: boolean = false): Promise<void> 
         }
       }
     })
+
+    // after migration complete, migrate "_configuration;:quickNotes"
+    migrateQuickNotes()
 
     await CommandBar.prompt('Template Migration', `${newNoteCounter} Templates Converted Successfully`)
   } catch (error) {